v3/requirements.txt
# Suntory v3 - Requirements
# Simplified but production-grade multi-agent system

# AutoGen Core (latest stable)
autogen-agentchat>=0.7.0
autogen-core>=0.7.0
autogen-ext>=0.7.0

# LLM Provider Abstraction (multi-provider support)
litellm>=1.51.0

# Configuration Management
pydantic>=2.11.0
pydantic-settings>=2.7.0
python-dotenv>=1.1.0
pyyaml>=6.0.0

# Vector Database for Agent Memory
chromadb>=0.5.23
sentence-transformers>=3.3.1

# Database (Async)
sqlalchemy>=2.0.36
aiosqlite>=0.20.0

# Rich Terminal UI
rich>=14.0.0
textual>=1.0.1
prompt-toolkit>=3.0.48

# HTTP Client
httpx>=0.28.0

# OpenAI & Anthropic APIs
openai>=1.96.0
anthropic>=0.43.0
google-generativeai>=0.8.3

# OpenTelemetry (Observability)
opentelemetry-api>=1.35.0
opentelemetry-sdk>=1.35.0
opentelemetry-instrumentation>=0.51b0

# Magentic One Dependencies
playwright>=1.49.0
beautifulsoup4>=4.12.3
lxml>=5.3.0

# Utilities
tenacity>=9.1.0
structlog>=24.4.0
click>=8.1.8
psutil>=6.1.1

# Testing
pytest>=8.3.0
pytest-asyncio>=0.24.0
pytest-cov>=6.0.0
pytest-timeout>=2.3.1

# Docker SDK (for sandboxing)
docker>=7.1.0
